Add files via upload
This code explains a critical programming interview problem: converting an integer number to a roman number. We will understand the approach in 2 essential steps. The first step is to derive the required roman notations from the given notations. The second step is to know how to break a number in the most feasible way required by the problem. This problem is among the most frequent programming interview problems and has been frequently asked for Amazon interview.